Remove compatibility methods at hosttracker-impl
Three List type variables have been converted to Map type at file
Host.java.

One BigInteger type variable has been converted to Uint64 type at
file SimpleAddressObserver.java. Mathematical operations using a
BigInteger variable are still used because there are no equivalent
methods for Uint64.

+++ hosttracker/implementation/src/main/java/org/opendaylight/l2switch/hosttracker/plugin/internal/SimpleAddressObserver.java
@@ -32,6 +32,7 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.packet.ipv4.rev140528.ipv4.
 import org.opendaylight.yang.gen.v1.urn.opendaylight.packet.ipv6.rev140528.Ipv6PacketListener;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.packet.ipv6.rev140528.Ipv6PacketReceived;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.packet.ipv6.rev140528.ipv6.packet.received.packet.chain.packet.Ipv6Packet;
+import org.opendaylight.yangtools.yang.common.Uint64;
 
 /**
  * A Simple Address Observer based on l2switch address observer.
@@ -187,7 +188,7 @@ public class SimpleAddressObserver implements ArpPacketListener, Ipv4PacketListe
          */
         BigInteger id = BigInteger.valueOf(ketype.getIntValue()).abs()
                 .add(BigInteger.valueOf(srcMacAddr.hashCode()).abs().shiftLeft(16));
-        addrs.setId(id);
+        addrs.setId(Uint64.valueOf(id));
         addrs.withKey(new AddressesKey(addrs.getId()));
         addrs.setVlan(vlanId);
         addrs.setIp(srcIpAddr);
